requirements.txt
python-telegram-bot==21.10
yt-dlp==2026.3.17
